Add K8sTokenProvider
This is a PoC to grab a k8s token from a kube config.

It doesn't work for certificates, and currently can not be modified to
work for an exec plugin returning a certificate, as the k8s java client
does not support it.

Issue: https://github.com/cloudfoundry/cf-crd-explorations/issues/80